+++ src/main/java/com/belhard/bookstore/dao/user/UserDaoImpl.java
@@ -98,7 +98,7 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public User delete(int id) {
+    public User delete(Long id) {
         try (Connection connection = dataSource.getConnection()) {
             logger.debug("Deleting user", id);
             PreparedStatement statement = connection.prepareStatement(DELETE_QUERY);
